As regras de estilo foram migradas para um modulo css (estavam escritas como objetos em .jsx).
