+++ src/components/VaDeCaronae/colors.jsx
@@ -1,0 +1,14 @@
+const colors = {
+  background1: 'rgb(43, 56, 138)',
+  background2: 'rgb(50, 64, 153)',
+  background3: 'rgb(62, 75, 154)',
+  text: 'rgb(254, 254, 254)',
+  highlight: 'rgb(255, 255, 255)',
+  shadow: 'rgb(23, 23, 23)',
+  peopleIconBackground: 'rgb(235, 114, 172)',
+  shieldIconBackground: 'rgb(110, 54, 31)',
+  leafIconBackground: 'rgb(23, 178, 112)',
+  jeepIconBackground: 'rgb(237, 128, 36)'
+};
+
+export { colors };
